08.12.2024 Исправление замечаний по классу InMemoryHistoryManager
