--- src/manager/Node.java
@@ -1,13 +1,0 @@
-package manager;
-
-public class Node<E> {
-    public E data;
-    public Node<E> next;
-    public Node<E> prev;
-
-    public Node(Node<E> prev, E data, Node<E> next) {
-        this.data = data;
-        this.next = next;
-        this.prev = prev;
-    }
-}
